requirements.txt
Flask
flask-cors
Flask-MySQLdb
bleach